+++ docs/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## ZSDK v1.0.1 (In Progress) [@Ryan Ulrick](mailto:rulrick@zscaler.com)
 - Update README.md with appropriate doc links. Iterate pyproject.toml to minor rev 1
 
+## ZSDK v1.1.0 (8-11-2023) [@Ryan Ulrick]
+- Updated logging handlers
+
 ## ZSDK v1.0.0 (8-11-2023) [@Ryan Ulrick](mailto:rulrick@zscaler.com)
 - Feat: All ZIA methods
 - Feat: Tests placeholders for all intermediate_ca_certs.intermediate_ca_certificates sub-methods